requirements.txt
# PiCor Requirements
# Generated based on actual installation process

# Core ML/RL libraries
torch==2.6.0+cu118
torchvision==0.21.0+cu118
torchaudio==2.6.0+cu118
torchrl==0.7.2
tensordict==0.7.0

# Gym and environment libraries
gym==0.19.0
mujoco_py==2.0.2.8
pybullet==3.2.7
termcolor==3.1.0

# MetaWorld (specific commit)
git+https://github.com/Farama-Foundation/Metaworld.git@04be337a12305e393c0caf0cbf5ec7755c7c8feb

# Garage (with no-deps to avoid conflicts)
garage==2021.3.0

# Build tools (specific versions for compatibility)
Cython==0.29.21
setuptools==59.5.0
wheel==0.38.0

# Core dependencies
numpy==1.23.5
scikit-image==0.24.0
scipy==1.13.1

# Utility libraries
akro==0.0.8
dowel==0.0.3
psutil==7.0.0
cma==2.7.0
setproctitle>=1.0

# TensorFlow ecosystem
tensorflow==2.19.0
tensorflow-probability>=0.11.0

# Ray for distributed computing
ray==2.10.0

# Logging and monitoring
wandb==0.21.0
tensorboard==2.19.0
tensorboardX==2.6.4

# Configuration and utilities
hydra-core==0.11.3
python-dotenv==1.1.1
PyYAML==6.0.2

# Type hints
typing_extensions==4.14.1

# Other dependencies
tqdm==4.67.1
requests==2.32.4
pillow==11.3.0
matplotlib==3.9.4

# Local custom libraries (to be installed separately)
# -e ./libs/custom_dmcontrol
# -e ./libs/custom_dmc2gym
